% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Character_functions.R
\name{create_character}
\alias{create_character}
\title{Main function to create the character}
\usage{
create_character()
}
\value{
An object of class \code{gamecharacter}
}
\description{
Creates Character with Name race and class, in addition other functions
are used to determine the characters stats.
}
\examples{
hero <- create_character()
}
