% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Character_functions.R
\name{Determine_stats}
\alias{Determine_stats}
\title{Initial stat function}
\usage{
Determine_stats(race, class)
}
\arguments{
\item{class}{string}

\item{obj}{A gamecharacter object.}
}
\value{
A list which holds the stats
}
\description{
The function is only used in create_character function, to determine
initial stats.
}
