% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Character_functions.R
\name{level_XP_update}
\alias{level_XP_update}
\title{Update XP and Level of a Game Character}
\usage{
level_XP_update(obj, gained_xp)
}
\arguments{
\item{obj}{A \code{gamecharacter} object.}

\item{gained_xp}{Numeric value. The amount of XP gained.}
}
\value{
An updated \code{gamecharacter} object with adjusted XP and level.
}
\description{
This function updates a game character's XP by adding the gained amount and checks
if the character should level up. If XP exceeds 100, the level increases and
XP is reset to the remainder.
}
