% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Character_functions.R
\name{print.gamecharacter}
\alias{print.gamecharacter}
\title{Print method for gamecharacter}
\usage{
\method{print}{gamecharacter}(x, ...)
}
\arguments{
\item{x}{A gamecharacter object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the object.
}
\description{
Print method for gamecharacter
}
\examples{
print(my_hero)
}
