% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Character_functions.R
\name{summary.gamecharacter}
\alias{summary.gamecharacter}
\title{Character summary function}
\usage{
\method{summary}{gamecharacter}(obj)
}
\arguments{
\item{obj}{}
}
\value{
the summary for the character
}
\description{
Character summary function
}
\examples{
Character Summary
===========================
Name      | Keld
Race      | Human
Class     | Mage
Level     | 1
HP        | 104
Strength  | 2
Intellect | 6
===========================
}
