% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Character_functions.R
\name{Update_stats}
\alias{Update_stats}
\title{Update character stats using available points}
\usage{
Update_stats(obj, choices = NULL)
}
\arguments{
\item{obj}{A gamecharacter object.}
}
\value{
The updated character object.
}
\description{
Update character stats using available points
}
\examples{
my_hero <- Update_stats(my_hero)
}
